requirements.txt
asgiref==3.3.4
Django==3.0.5
djangorestframework==3.12.4
djongo==1.3.4
dnspython==2.1.0
gunicorn==20.1.0
mongoengine==0.23.0
pymongo==3.11.3
pytz==2021.1
sqlparse==0.2.4
typing-extensions==3.7.4.3
